Refactor PeerLogin and PeerSync types to use the new types package
